Vectorize the ChaCha20 when building on AMD64.
This is the entirely wrong way to write optimized crypto in Go, but
till they either add SIMD intrinsics, or switch to an assembly language
format that isn't special snowflake bullshit, this will have to suffice.

AFAIK there's one other person that used this approach, but when I
looked at their implementation, it was oneshot only, and furthermore
will SIGSEGV when it encounters unaligned data.

(NB: I assume your cgo compiler has stdint.h.  It was introduced in
 1999 as part of the standard, there is absolutely no excuse.)

For the large benchmark this is a hair under 3x increase in throughput.